+++ app/controllers/admin/users_controller.rb
@@ -44,7 +44,7 @@ class Admin::UsersController < ApplicationController
   def user_params
     params.require(:user).permit(
       :username, :full_name, :email, :approved, :admin, :editor, :role,
-      :key_contact, :phone, :keywords
+      :key_contact, :phone, :keywords, :site_id
     )
   end