+++ db/migrate/20180815182521_add_site_id_to_users.rb
@@ -1,0 +1,6 @@
+class AddSiteIdToUsers < ActiveRecord::Migration[5.2]
+  def change
+    add_column :users, :site_id, :integer
+    add_index :users, :site_id
+  end
+end